Input validation in class Request
